requirements.txt
pandas>=2.0.0
numpy>=1.20.0
python-dotenv>=1.0.0
langchain>=0.1.0
sentence-transformers>=2.2.2
chromadb>=0.4.18
tqdm>=4.65.0
google-generativeai>=0.3.0
pathlib>=1.0.1
